Bugfix - 'Disable encoding on preview'
Added the default state for 'disable encoding while previewing'. Before
when you hit 'defaults' in advanced with test build v0.52.04b it would
not disable that option or do anything at all, clearly not intended. It
now properly unchecks this option which is also how it is set by
default.