Add DirectShow device list to the plugin
Add a list with known device IDs to the DirectShow plugin. An attempt to
work around how Windows sometimes switches the order around for
installed capture devices, which could cause devices of the same model
to stop working once reconfigured. Virtual DirectShow devices are
unaffected by this, as they have no device ID to begin with.