Add nvidia optimus hint if user selects nvidia
Right now, the state of things is that if the user was running an nvidia
laptop, he/she "had" to use the first adapter, and then fool with the
nvidia control panel to get the nvidia working properly.

This change makes it so that the nvidia optimus hint will be activated
and the first adapter will be forced if the user selects the nvidia
adapter in OBS' video adapter drop down, making it much easier for users
to deal with the nvidia laptop issue.