Add UI for log file uploader
Also includes UI for opening specific log files directly from the OBS UI

On successful upload the resulting (non-API) URL is copied into the
clipboard and a message box signalling successful upload is displayed.
This should probably be expanded into a dialog with buttons to retry
copying the URL to the clipboard and maybe a text box with the URL
preselected.

Whether or not the upload URL should be cached (per log file) and reused
in case of duplicate upload (attempts) is still open for debate.