Fix FLV writer
The first FLV packet was getting thrown off by initialTimeStamp, because
I foolishly forgot that even if the function was called with 0 as the
argument for the timestamp, it would still subtract the -1 from that.
Also made the capitalization of timestamp to be consistent with other
variable in the file.