Fix AAC mono output
The entire mono situation is my fault.  I did test it and it seemed
like it was working at first when I got the pull request so I said "hey
okay looks good."  Unfortunately I was mistaken, so I'm fixing the code
myself.

AAC was taking in stereo audio and was not calculating the timestamps
correctly.  Internally OBS still gets data in the form of stereo, so
because this is OBS1 I'm just going to put in this workaround code that
downmixes stereo to mono in to the AAC encoder buffer.