Fix JSON parsing for XConfig
This parser really wasn't originally meant for JSON, it was originally
meant to be a form of HJSON, but this change should allow it to parse
both JSON and OBS' HJSON format.  It simply detects whether it starts
with an empty element (like JSON always does) and then it treats it as
JSON if present.  OBS' format doesn't use the empty element.