Revert to old hooks for the time being
Minhooks is breaking for too many people right now, so I'm removing it
until further testing and verification.

This just goes to show, you can test for weeks and still not get
comprehensive feedback.

Also added some light obfuscation of SetWindowsHookEx and
UnhookWindowsHookEx to prevent issues with game capture and A/V.