Fix game capture hotkey and LoL capturing
This fixes league of legends capturing issues where people would get a
black screen with game capture with it:  What happened is that riot must
have changed the window class for the hidden patcher window to have the
same window class as the in-game window, so when you exit your game, it
would try to capture the hidden patcher window and it would get stuck on
it.  Solution was to make it so only visible windows were captured.

Secondly, this improves the game capture hotkey feature.  Game capture
hotkey will now always 'stick' to the last window you captured, and save
it to config.  Only until the hotkey is hit again will it try to capture
a different window.

Windows that are not visible will no longer be used as capture targets.
I'm guessing that this may have been the cause of a number of prior
issues.