Implement scene collections in menu only
This hopefully resolves some conflicting opinions about how scene
collections should be handled, and removes it from the general settings
panel, where it felt a bit cluttered.

Actually, I would kind of like it if profiles were handled this way too.
This feels a lot cleaner than how profiles are handled in the general
settings.  The profiles thing in the general settings definitely feels
clunky.

In the main menu, the scene collections menu now has the following menu
items:

- New
- Duplicate
- Rename
- Remove
- Import
- Export

Duplicate was something I personally added - I felt like it might be a
nice thing to have for the sake of convenience.  Import/Export was
requested by dodgepong, and is also definitely nice to have.  Due to how
easy new scene collections are to set up it didn't take very long to put
in.