Keep Global Sources in "sceneCollection/scenes.xconfig".
Instead of each Scene Collection having its own Global Sources. They are stored in and read from "sceneCollection/scenes.xconfig".

With this change users won't need to recreate their Global Sources for every new Scene Collection.

When a user deletes a Global Source it is removed from all Scene Collections.

The only issue I know is when a user renames a Global Source the user will need to add it back to the scene.