Added "Copy To" in the scenes right click context menu
I have added a "Copy To" menu item to the scenes right click context menu below "Copy".

When hovered over it will open a sub-menu where the user can then select the scene collection file the user wants to copy the selected scene to.

If there is a reference to a global source it will open a dialog asking the user if he/she wants to also copy the global sources the references point to.If the user clicks "No" the references are removed from the copied scene.If the global sources already exists in the target scene collection a dialog will open letting the user know but the scene is still copied.

A dialog will open telling the user the export was successful or the scene already exists in the selected scene collection.

I believe this is useful for users that want to copy or move scenes to a different scene collection.