Update services.xconfig
The UK Ingest was changed into EU-North. (Where it was in the first place) and EU-North (Old) was live.ams.hitbox.tv which still works, but is not advertised in their API so I've removed it. I also decided not to add Asia as they have the ingest url marked as US_West. Which is obviously a mistake.

Changes were updated via their API. http://i.imgur.com/lwvM6tL.png (Image because the Ingest API requires auth.)