Fixed a bug with "copyTo" that only effected the 64bit version of obs
I have fixed a bug when a user uses obs 64bit and  copies a scene to a
different scene collection it will instead of copying to the target
scene collection it would copy it to a new scene collection with what
looks like chinese characters as the new scene collection's name. This
bug only happened in the 64bit version of obs.