Rename Live Chat to Community Chat and update URL
The Live Chat got renamed to Community Chat on the main website to make
more clear that the helpers in the chat are all volunteers.

Also the URL got updated and does now point to a site with instructions
for the chat instead of pointing directly to the QuakeNet webchat.