Don't snap crop when resizing a source
When resizing a source with snapping enabled, any crop that was small enough to
be snapped would be reset as if the user was dragging the crop line with snap
enabled. The crop should not be affected by a resize action (unless the source
becomes too small), so disable the snap check on crops when resizing.