Build: Mark manifest as DPI-aware
Even though OBS isn't DPI-aware, we need to pretend we are or Windows will enable DPI virtualization. This makes it so that when the user first runs OBS, we get an incorrect base resolution reported which causes "My X source is cut off" problems down the line.

DPI virtualization seems incompatible with BitBlt and likely some other drawing functions in the first place so best not to use it.