DShowPlugin: Only dynamically change size in Preprocess
newCX and newCY were being changed when a sample was received.  However,
the sample associated with that size change will not be processed until
it's used in Preprocess, so it could often change the texture size
before the sample was even used.  This could cause crashes under certain
circumstances.

A simple fix to this is to only change newCX/newCY when the sample is
actually used.