NVENC: Always use 2PASS_QUALITY rate control
It does a very decent job at holding a constant bitrate now and has a major
impact on the visual quality. Using the inferior rc mode when CBR padding
is enabled, which propably is the cast almost everywhere, seems like a bad
decision to me.
I wasn't able to spot a notable difference in the quality of the constantness
these two provide, so just removing the framesize capped one.